[FIX] Rename :core:design-system to :core:designsystem
