+++ build-logic/convention/src/main/kotlin/fitrun.android.feature.gradle.kts
@@ -27,6 +27,6 @@ dependencies {
     implementation(findLibrary("orbit.viewmodel"))
     implementation(findLibrary("orbit.compose"))
 
-    implementation(project(path = ":core:design-system"))
+    implementation(project(path = ":core:designsystem"))
     implementation(project(path = ":core:common"))
 }
