+++ core/designsystem/src/main/kotlin/com/yapp/fitrun/core/designsystem/Font.kt
@@ -6,7 +6,6 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import com.yapp.fitrun.core.design_system.R
 
 val pretendardFamily = FontFamily(
     Font(R.font.pretendard_bold, FontWeight.Bold, FontStyle.Normal),
