+++ core/designsystem/src/main/res/drawable/logo_splash_text.xml
