+++ core/designsystem/src/main/res/font/pretendard_medium.otf
