+++ core/designsystem/src/main/res/font/pretendard_regular.otf
