+++ core/designsystem/src/main/res/font/pretendard_semi_bold.otf
