+++ core/designsystem/src/main/res/mipmap-anydpi/ic_launcher_round.xml
