+++ core/designsystem/src/main/res/mipmap-mdpi/ic_launcher.webp
