+++ core/designsystem/src/main/res/mipmap-xhdpi/ic_launcher.webp
