+++ feature/splash/src/main/kotlin/com/yapp/fitrun/feature/splash/FitRunSplashScreen.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.yapp.fitrun.feature.splash.viewmodel.FitRunSplashViewModel
-import com.yapp.fitrun.core.design_system.R
+import com.yapp.fitrun.core.designsystem.R
 import com.yapp.fitrun.core.designsystem.FitRunOrange
 import com.yapp.fitrun.feature.splash.viewmodel.FitRunSplashSideEffect
 import org.orbitmvi.orbit.compose.collectAsState
